Guard against null render data in renderables
Over the last few months, there have been a few reports of
NullPointerExceptions in LegendRenderable nad RadarRenderable.  I don't
have enough context to know why the value is null (it shouldn't be).
This change checks if it is null and skips the render if it is, should
prevent those crashes.